+++ internal/deployer/deployer_api.go
@@ -1,4 +1,4 @@
-package swarm
+package deployer
 
 import (
 	"context"
