+++ internal/deployer/deployer_test.go
@@ -1,4 +1,4 @@
-package swarm
+package deployer
 
 import (
 	"strconv"
