+++ internal/deployer/runner.go
@@ -1,0 +1,7 @@
+package deployer
+
+import "context"
+
+type Runner interface {
+	Run(ctx context.Context, command string, args ...string) (string, error)
+}
